msm: pm: Add API to enable/disable retention mode On 8974 targets, retention mode voltage is controlled by the LDO regulator, which in turn is powered through a ganged PMIC regulators. At certain low voltage levels, the ganged PMIC will not be able to power the LDO retention voltage. The regulator framework would invoke this API to enable/disable retention mode in these scenarios.
When disabling retention modes, send a IPI to wakeup all online cores
out of idle, to ensure that no core is in a retention low power mode.
Also, disabled support for retention during suspend/hotplug operations
to ensure that a CPU isn't already in retention when the voltage is
lowered. This shouldn't have an effect on suspend, as the driver always
chooses the deepest sleep mode for suspend/hotplug.

Change-Id: I0af18168968c2aaf05ca99188d30762612c87de6
Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>